Http server

+++ src/server/database/UserDB.java
@@ -1,12 +1,14 @@
 package server.database;
 
 import server.enums.error.ErrorType;
+import server.user.User;
 
 import java.sql.Array;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 
 public class UserDB extends SQLDB {
     public static void main(String[] args) {
@@ -22,28 +24,29 @@ public class UserDB extends SQLDB {
             preparedStatement.setString(1, username);
             preparedStatement.setString(2, password);
             ResultSet resultSet = preparedStatement.executeQuery();
-            if(resultSet.next()){
+            if (resultSet.next()) {
                 StringBuffer stringBuffer = new StringBuffer();
                 stringBuffer.append(resultSet.getInt("id"));
                 return stringBuffer.toString();
             }
             return null;
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            System.out.println(e.getMessage());
+            return null;
         }
     }
 
 
-    public static int createUser(String username , String password){
-        try{
+    public static int createUser(String username, String password) {
+        try {
             preparedStatement = connection.prepareStatement("INSERT INTO users (username , password) VALUES (?,?) returning id");
-            preparedStatement.setString(1,username);
-            preparedStatement.setString(2,password);
+            preparedStatement.setString(1, username);
+            preparedStatement.setString(2, password);
 
             ResultSet resultSet = preparedStatement.executeQuery();
             resultSet.next();
             return resultSet.getInt("id");
-        }catch (SQLException e){
+        } catch (SQLException e) {
             System.out.println(e.getMessage());
             throw new RuntimeException(e);
         }
@@ -65,11 +68,11 @@ public class UserDB extends SQLDB {
         SQLDB.updateFieldsKeys("users", userId, userUpdate);
     }
 
-    //TODO Int id , userId?;
     public static ErrorType follow(int userId, int targetId) {
+        if(userId == targetId) return ErrorType.SAME_PERSON;
+        if (!SQLDB.containFieldKey("users", "id", targetId)) return ErrorType.DOESNT_EXIST;
         if (SQLDB.containInArrayFieldObject("users", userId, "following", targetId)) return ErrorType.ALREADY_EXIST;
         if (SQLDB.containInArrayFieldObject("users", targetId, "blocked", userId)) return ErrorType.BLOCKED;
-        //TODO check for reputation?
 
 
         SQLDB.appendToArrayField("users", userId, "following", targetId);
@@ -77,23 +80,30 @@ public class UserDB extends SQLDB {
         return ErrorType.SUCCESS;
     }
 
-    public static void unFollow(int userId, int targetId) {
+    public static ErrorType unFollow(int userId, int targetId) {
+        if(userId == targetId) return ErrorType.SAME_PERSON;
+        if (!SQLDB.containFieldKey("users", "id", targetId)) return ErrorType.DOESNT_EXIST;
+        if (!SQLDB.containInArrayFieldObject("users", userId, "following", targetId)) return ErrorType.HAVE_NOT_FOLLOWED;
         SQLDB.removeFromArrayField("users", userId, "following", targetId);
         SQLDB.removeFromArrayField("users", targetId, "follower", userId);
 
+        return ErrorType.SUCCESS;
     }
 
 
     public static Array getFollower(int userId) {
         return (Array) getFieldObject("users", userId, "follower");
-    }
+    }//TODO handle null
+
     public static Array getFollowing(int userId) {
         return (Array) SQLDB.getFieldObject("users", userId, "following");
     }
 
     public static ErrorType block(int userId, int targetId) {
-        if (isBlocked(targetId, userId)) return ErrorType.ALREADY_EXIST;
-        if (SQLDB.containInArrayFieldObject("users" , userId , "following" , targetId)){
+        if(userId == targetId) return ErrorType.SAME_PERSON;
+        if (!SQLDB.containFieldKey("users", "id", targetId)) return ErrorType.DOESNT_EXIST;
+        if (isBlocked(userId, targetId)) return ErrorType.ALREADY_EXIST;
+        if (SQLDB.containInArrayFieldObject("users", userId, "following", targetId)) {
             SQLDB.removeFromArrayField("users", userId, "following", targetId);
             SQLDB.removeFromArrayField("users", targetId, "follower", userId);
         }
@@ -101,15 +111,53 @@ public class UserDB extends SQLDB {
         return ErrorType.SUCCESS;
     }
 
-    public static void unBlock (int userId, int targetId) {
+    public static ErrorType unBlock(int userId, int targetId) {
+        if(userId == targetId) return ErrorType.SAME_PERSON;
+        if (!SQLDB.containFieldKey("users", "id", targetId)) return ErrorType.DOESNT_EXIST;
+        if (!SQLDB.containInArrayFieldObject("users", userId, "blocked", targetId)) return ErrorType.HAVE_NOT_BLOCKED;
         SQLDB.removeFromArrayField("users", userId, "blocked", targetId);
+        return ErrorType.SUCCESS;
     }
 
-    public static boolean isBlocked(int userId, int targetId){
+    public static boolean isBlocked(int userId, int targetId) {
         return containInArrayFieldObject("users", userId, "blocked", targetId);
     }
 
-    public static ArrayList<Object> getUsersId(){
+    public static ArrayList<Object> getUsersId() {
         return getAllKeyOfField("users", "id");
     }
+
+    public static User getUser(int targetId) {
+        try {
+            ResultSet resultSet = getResultSet("users", targetId);
+            if (!resultSet.next()) return null;
+
+            String username = resultSet.getString("username");
+            Object[] followerId = null;
+            Object[] followingId = null;
+            HashSet<Integer> follower = new HashSet<>();
+            HashSet<Integer> following = new HashSet<>();
+
+            if (resultSet.getArray("follower" )!= null){
+                followerId = (Object[]) (resultSet.getArray("follower").getArray());
+                for (Object tmp : followerId) {
+                    follower.add((int) tmp);
+                }
+            }
+            if (resultSet.getArray("following") != null){
+                followingId = (Object[]) (resultSet.getArray("following").getArray());
+                for (Object tmp : followingId) {
+                    following.add((int) tmp);
+                }
+
+            }
+
+            User user = new User(targetId, username, following, follower);
+            return user;
+
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
 }
